embedded - sketch fix - float angle

+++ embedded/src/sketch_blocking/sketch_blocking.ino
@@ -9,7 +9,7 @@
 
 
 char buffer2serialData[CONST_UART_BUFFER_SIZE];
-byte angle_base, angle_mid, angle_hand;
+float angle_base, angle_mid, angle_hand;
 ServoMotor base = ServoMotor(90, 0, 180, 10, -5);
 ServoMotor mid = ServoMotor(90, 0, 90, 20, 10);
 ServoMotor hand = ServoMotor(90, 35, 180);
@@ -53,13 +53,13 @@ void uartLoop(void(*cb)()) {
 
 void handleUartReceivedData() {
   char *token = strtok(buffer2serialData, ",");
-  angle_base = atoi(token);
+  angle_base = atof(token);
 
   token = strtok(NULL, ",");
-  angle_mid = atoi(token);
+  angle_mid = atof(token);
   
   token = strtok(NULL, ",");
-  angle_hand = atoi(token);
+  angle_hand = atof(token);
 
   base.Write(angle_base);
   mid.Write(angle_mid);
